moved application classes to their package to improve archetype generation
