+++ java/src/main/java/com/amido/stacks/app/ApplicationNoSecurity.java
@@ -1,4 +1,4 @@
-package com.amido.stacks;
+package com.amido.stacks.app;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
